+++ packages/core/core-api/src/facet/index.ts
@@ -34,6 +34,7 @@ export type {
   IgnoreFileRole,
   LicenseRole,
   PackageFileRole,
+  PackageManager,
   PackageManagerRole,
   ScriptExplanation,
   VersionControlRole,